Desmantelado de la clase gestor en clases por funcionalidades
