+++ src/dominio/Cambiar_Precio_Zapato.java
@@ -1,0 +1,19 @@
+package dominio;
+
+import java.util.stream.Collectors;
+
+public class Cambiar_Precio_Zapato {
+
+    private GestorZapatos miGestorZapatos;
+
+    private Cambiar_Precio_Zapato(GestorZapatos gz){
+        this.miGestorZapatos = gz;
+    }
+
+    public boolean cambiarPrecio(String codigo, double neuvo,int dia,int mes,int anio) {
+		Zapato aux;
+		aux = miGestorZapatos.getZapatos().stream().filter(p -> p.getCodigo().compareTo(codigo)==0).collect(Collectors.toList()).get(0);
+		aux.setPrecio(neuvo, dia, mes, anio);
+		return true;
+	}
+}
